+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## Unreleased
-### Changed
+
+# 2.0.13
 * update db_migrate.sh
 * update rails log file mode
 * add user `consul`  to group `ubuntu`
+* complete db migrate task
 
 # 2.0.12
 * fix td-agent restart service